+++ client/src/config/setsSearchTags.ts
@@ -1,0 +1,7 @@
+import { IHalLink } from '../types/IHalLink'
+
+// Used to request and display unit data
+export const archive: IHalLink = {
+  title: 'View records from this archive with images',
+  searchTag: 'lux:setItemsWithImages',
+}